:mag: nit: break apart noa metric alert handler tests
